+++ src/kernel/SigmoidKernel.cpp
@@ -21,7 +21,7 @@ extern "C" {
 }
 #endif
 
-CSigmoidKernel::CSigmoidKernel(LONG size, DREAL g, DREAL c)
+CSigmoidKernel::CSigmoidKernel(INT size, DREAL g, DREAL c)
   : CSimpleKernel<DREAL>(size),gamma(g), coef0(c)
 {
 }
